Refactor existing dft methods to separate modules
